backend/requirements.txt
FastAPI==0.95.0
SQLAlchemy==2.0.0
asyncpg==0.27.0
pydantic==1.10.7
uvicorn==0.22.0
pytest==7.2.2
pytest-asyncio==0.20.2
httpx==0.23.3
psycopg2
python-dotenv